Move few method from layer/geoserver/__init__.py to util.py
