+++ src/layman/layer/geoserver/__init__.py
@@ -3,14 +3,10 @@ from dataclasses import dataclass
 
 from flask import g
 
-import crs as crs_def
 from geoserver import util as gs_util
 from layman.http import LaymanError
-from layman import settings, util as layman_util
-from layman.common import bbox as bbox_util, geoserver as gs_common, empty_method
-from . import wms
-from .util import get_external_db_store_name
-from ..layer_class import Layer
+from layman import settings
+from layman.common import empty_method
 
 logger = logging.getLogger(__name__)
 FLASK_RULES_KEY = f"{__name__}:RULES"
@@ -47,26 +43,6 @@ def delete_whole_user(username, auth=settings.LAYMAN_GS_AUTH):
 
 
 ensure_workspace = empty_method
-
-
-def create_external_db_store(workspace, *, uuid, table_uri, auth=settings.LAYMAN_GS_AUTH):
-    pg_conn = {
-        'host': table_uri.hostname,
-        'port': table_uri.port,
-        'dbname': table_uri.db_name,
-        'user': table_uri.username,
-        'password': table_uri.password,
-    }
-    store_name = get_external_db_store_name(uuid=uuid)
-    gs_util.create_db_store(workspace,
-                            auth,
-                            db_schema=table_uri.schema,
-                            pg_conn=pg_conn,
-                            name=store_name,
-                            )
-    return store_name
-
-
 delete_workspace = empty_method
 
 
@@ -92,43 +68,6 @@ def check_workspace_name(workspace):
     rolename = gs_util.username_to_rolename(workspace)
     if rolename in gs_util.RESERVED_ROLE_NAMES:
         raise LaymanError(35, {'reserved_by': __name__, 'role': rolename})
-
-
-def set_security_rules(*, layer: Layer, gs_names, access_rights, auth, ):
-    read_roles = access_rights.get('read') if access_rights and access_rights.get('read') else layer.access_rights['read']
-    write_roles = access_rights.get('write') if access_rights and access_rights.get('write') else layer.access_rights['write']
-
-    security_read_roles = gs_common.layman_users_and_roles_to_geoserver_roles(read_roles)
-    gs_util.ensure_layer_security_roles(gs_names.workspace, gs_names.name, security_read_roles, 'r', auth)
-
-    security_write_roles = gs_common.layman_users_and_roles_to_geoserver_roles(write_roles)
-    gs_util.ensure_layer_security_roles(gs_names.workspace, gs_names.name, security_write_roles, 'w', auth)
-
-
-def get_layer_bbox(*, layer: Layer):
-    # GeoServer is not working good with degradeted bbox
-    result = bbox_util.get_bbox_to_publish(layer.native_bounding_box, layer.native_crs)
-    return result
-
-
-def publish_layer_from_db(*, layer: Layer, gs_names, metadata_url, store_name=None):
-    bbox = get_layer_bbox(layer=layer)
-    lat_lon_bbox = bbox_util.transform(bbox, layer.native_crs, crs_def.EPSG_4326)
-    gs_util.post_feature_type(gs_names.workspace, gs_names.name, layer.description, layer.title, bbox, layer.native_crs, settings.LAYMAN_GS_AUTH, lat_lon_bbox=lat_lon_bbox, table_name=layer.table_uri.table, metadata_url=metadata_url, store_name=store_name)
-
-
-def publish_layer_from_qgis(*, layer: Layer, gs_names, metadata_url, ):
-    store_name = wms.get_qgis_store_name(uuid=layer.uuid)
-    info = layman_util.get_publication_info_by_class(layer, context={'keys': ['wms', ]})
-    layer_capabilities_url = info['_wms']['qgis_capabilities_url']
-    gs_util.create_wms_store(gs_names.workspace,
-                             settings.LAYMAN_GS_AUTH,
-                             store_name,
-                             layer_capabilities_url)
-    bbox = get_layer_bbox(layer=layer)
-    lat_lon_bbox = bbox_util.transform(bbox, layer.native_crs, crs_def.EPSG_4326)
-    gs_util.post_wms_layer(gs_names.workspace, gs_names.name, layer.qgis_names.name, store_name, layer.title, layer.description, bbox, layer.native_crs, settings.LAYMAN_GS_AUTH,
-                           lat_lon_bbox=lat_lon_bbox, metadata_url=metadata_url)
 
 
 def get_usernames():
