+++ src/layman/layer/geoserver/wfs.py
@@ -86,7 +86,7 @@ def get_wfs_direct():
         settings.LAYMAN_GS_AUTHN_HTTP_HEADER_ATTRIBUTE: settings.LAYMAN_GS_USER,
     }
     ows_url = get_wfs_url()
-    from .util import wfs_direct
+    from geoserver.util import wfs_direct
     key = get_flask_proxy_key()
     redis_obj = settings.LAYMAN_REDIS.hgetall(key)
     string_value = redis_obj['value'] if redis_obj else None
