+++ src/layman/layer/geoserver/wfs_tasks.py
@@ -3,8 +3,7 @@ from layman import celery_app, settings
 from layman.celery import AbortedException
 from layman.layer.layer_class import Layer
 from . import wfs
-from .util import get_db_store_name
-from .. import geoserver
+from .util import get_db_store_name, get_layer_bbox
 
 headers_json = gs_util.headers_json
 
@@ -29,7 +28,7 @@ def patch_after_feature_change(
     if layer_data.geodata_type != settings.GEODATA_TYPE_VECTOR:
         raise NotImplementedError(f"Unknown geodata type: {layer_data.geodata_type}")
 
-    bbox = geoserver.get_layer_bbox(layer=layer_data)
+    bbox = get_layer_bbox(layer=layer_data)
     store_name = get_db_store_name(uuid=layer_data.uuid, original_data_source=layer_data.original_data_source.value)
     gs_layername = layer_data.gs_names.wfs
     gs_util.patch_feature_type(gs_layername.workspace, gs_layername.name, auth=settings.LAYMAN_GS_AUTH, bbox=bbox, crs=layer_data.native_crs,
