+++ src/layman/layer/geoserver/wms_tasks.py
@@ -5,8 +5,7 @@ from layman.common import bbox as bbox_util
 from layman.celery import AbortedException
 from layman.layer.layer_class import Layer
 from . import wms
-from .util import get_db_store_name
-from .. import geoserver
+from .util import get_db_store_name, get_layer_bbox
 
 headers_json = gs_util.headers_json
 
@@ -27,7 +26,7 @@ def patch_after_feature_change(
     layer_data = Layer(layer_tuple=(workspace, layer))
 
     if layer_data.geodata_type == settings.GEODATA_TYPE_VECTOR:
-        bbox = geoserver.get_layer_bbox(layer=layer_data)
+        bbox = get_layer_bbox(layer=layer_data)
         wms_layername = layer_data.gs_names.wms
         lat_lon_bbox = bbox_util.transform(bbox, layer_data.native_crs, crs_def.EPSG_4326)
         if layer_data.style_type == 'sld':
